+++ wies/core/static/js/dialog.js
@@ -44,6 +44,7 @@ document.addEventListener("closeModal", function () {
   const modalContainers = [
     "labelFormModal",
     "userFormModal",
+    "suborganizationFormModal",
     "clientModalContainer",
   ];
   modalContainers.forEach((modalId) => {